ice4hpc/xAMM/ice4hpc/src/requirements.txt
ansible==2.9.27
apipkg==1.5
argcomplete==1.9.3
asn1crypto==0.24.0
astroid==2.3.3
attrs==17.4.0
Babel==2.5.1
backoff==1.11.1
bcc==0.25.0
bcrypt==3.1.6
beautifulsoup4==4.6.3
bitstring==3.1.7
blivet==3.6.0
Bottleneck==1.2.1
Brlapi==0.6.7
cachetools==3.1.1
CCColUtils==1.5
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
click==6.7
cloud-init==23.4
clufter==0.77.1
ClusterShell==1.9.2
colorama==0.4.3
configobj==5.0.6
configshell-fb==1.1.28
coverage==4.5.1
createrepo-c==0.17.7
createrepo-mod==0.1
cryptography==3.2.1
cssselect==0.9.2
cupshelpers==1.0
cycler==0.10.0
Cython==0.28.1
dasbus==1.2
dateutils==0.6.8
dbus-python==1.2.4
decorator==4.2.1
defusedxml==0.6.0
dir2module==0.1
distro==1.4.0
dnspython==1.15.0
docutils==0.14
dogpile.cache==0.6.8
ethtool==0.14
execnet==1.7.1
file-magic==0.3.0
Flask==0.12.2
future==0.18.3
gitdb==4.0.1
GitPython==3.1.18
google-auth==1.1.1
gpg==1.13.1
gssapi==1.5.1
html5lib==0.999999999
humanize==0.5.1
hvac==0.9.5
idna==2.5
importlib-metadata==0.23
include-server==3.3.3
iniparse==0.4
initial-setup==0.3.81.7
iotop==0.6
ipaclient==4.9.13
ipalib==4.9.13
ipaplatform==4.9.13
ipapython==4.9.13
isc==2.0
isort==4.3.21
itsdangerous==0.24
Jinja2==2.10.1
jmespath==0.9.0
jsonpatch==1.21
jsonpointer==1.10
jsonschema==2.6.0
jwcrypto==0.5.0
kerberos==1.3.0
kiwisolver==1.1.0
kmod==0.1
kobo==0.37.0
koji==1.34.0
kubernetes==11.0.0
langtable==0.0.51
lark-parser==0.8.1
lazy-object-proxy==0.0.0
libcomps==0.1.18
libgenders==1.2
libvirt-python==8.0.0
lit==17.0.6
lorax===devel
lxml==4.2.3
Mako==1.0.6.dev0
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==3.0.3
mccabe==0.6.1
meld==3.20.4
meld3==2.0.0
mod-wsgi==4.6.4
modulemd-merge==0.1
modulemd-tools==0.1
munkres==1.1.2
natsort==7.1.1
netaddr==0.7.19
netifaces==0.10.6
networkx==1.11
nftables==0.1
ntplib==0.3.3
numexpr==2.7.0
numpy==1.14.3
oauthlib==2.1.0
ordered-set==2.0.2
pandas==0.25.3
paramiko==2.12.0
pciutils==2.3.6
pcp==5.0
pcs==0.10.15
perf==0.1
pexpect==4.3.1
pid==2.1.1
Pillow==5.1.1
ply==3.9
podman-compose==1.0.6
prettytable==0.7.2
productmd==1.11
psutil==5.4.3
psycopg2==2.7.5
ptyprocess==0.5.2
pungi==4.1.38
pwquality==1.4.4
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycairo==1.16.3
pycparser==2.14
pycups==1.9.72
pycurl==7.43.0.2
pydbus==0.6.0
pyenchant==2.0.0
pygobject==3.28.3
pyinotify==0.9.6
PyJWT==1.6.1
pykickstart==3.16.16
pylint==2.4.4
PyNaCl==1.3.0
pyodbc===4.0.0-unsupported
pyOpenSSL==19.0.0
pyparsing==2.1.10
pyparted==3.11.7
pyperclip==1.8.2
PyQt5==5.15.0
PyQt5-sip==4.19.25
pyroute2==0.5.3
pyserial==3.1.1
PySocks==1.6.8
python-augeas==0.5.0
python-cpio==0.1
python-dateutil==2.6.1
python-dmidecode==3.12.2
python-dotenv==0.19.2
python-ldap==3.3.1
python-linux-procfs==0.7.3
python-meh==0.47.2
python-multilib==1.2
python-redfish==3.1.6
python-redfish-utilities==1.2.3
python-yubico==1.3.2
pytz==2017.2
pyudev==0.21.0
pyusb==1.0.0
pyxdg==0.25
PyYAML==3.12
qrcode==5.1
repo2module==0.1
requests==2.20.0
requests-cache==0.5.1
requests-file==1.4.3
requests-ftp==0.3.1
requests-futures==1.0.0
requests-gssapi==1.2.2
requests-kerberos==0.12.0
requests-mock==1.7.0
requests-oauthlib==1.0.0
requests-pkcs12==1.7
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
responses==0.10.5
rhnlib==2.8.6
rpkg==1.67
rpm==4.14.3
rsa==4.9
rtslib-fb==2.1.75
sasutils==0.5.0
schedutils==0.6
scipy==1.0.0
scons==3.1.2
scour==0.35
selinux==2.9
sepolicy==1.1
setools==4.3.0
setroubleshoot==1.1
sh==1.12.14
simpleline==1.1.3
six==1.11.0
slip==0.6.4
slip.dbus==0.6.4
smmap==3.0.1
sos==4.7.1
SSSDConfig==2.9.4
subscription-manager==1.28.42
suds-jurko==0.7.dev0
syspurpose==1.28.42
systemd-python==234
tables==3.5.2
targetcli-fb==2.1.53
templated-dictionary==1.4
tosspkg==1.3
tuna==0.18
typed-ast==1.4.0
typing-extensions==3.7.4.3
urllib3==1.24.2
urwid==1.3.1
virt-who==1.30.16
virtualenv==15.1.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.12.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==3.0.2
xxdiff-scripts==5.1
zipp==0.5.1
zope.event==4.2.0
zope.interface==4.6.0
